Fix 32-bit truncation of 64-bit module addresses in offset scanner
NetworkServer::current_offset returned UINT, truncating the QWORD
offset table (e.g. CharInfo at 0x140EB6440) before EQGameScanner
could use the full address as a base for structure-offset scanning.

findEQPointerOffset/findEQStructureOffset took and returned DWORD
despite handling full compile-time module addresses that exceed 32
bits, and callers were explicitly casting QWORD ini values down to
DWORD before passing them in. Widened both functions end-to-end to
QWORD and added proper 8-byte read support for typelen >= 8 (masks
with 8 't' characters previously fell through to a 4-byte DWORD
read, silently discarding the upper 32 bits).

+++ src/NetworkServer.cpp
@@ -339,9 +339,9 @@ void NetworkServer::enterReceiveLoop(MemReaderInterface* mr_intf)
 	}
 }
 
-UINT NetworkServer::current_offset(int type)
-{ 
-		return (UINT)offsets[(offset_types)type]; 
+QWORD NetworkServer::current_offset(int type)
+{
+		return offsets[(offset_types)type];
 }
 
 bool NetworkServer::processReceivedData(MemReaderInterface* mr_intf)
